pyproject.toml
[project]
name = "quantum-optical-bus-simulation"
version = "0.1.0"
description = "Hybrid quantum-classical calibration dashboard for squeezed-light TDM optical buses"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "scipy",
    "matplotlib",
    "strawberryfields",
    "streamlit",
]

[project.optional-dependencies]
# Meep requires conda or Linux
full = ["meep"]
demo = ["pillow"]
yaml = ["pyyaml"]
dev = [
    "ruff",
    "pre-commit",
]
test = [
    "pytest",
    "ruff",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    ".pytest_cache",
]

[tool.ruff.format]
line-ending = "lf"
